+++ frontend/components/gpi/Document.tsx
